Feature: Added functions that will generate a random question, will leave for others to implement into other parts of the code.
This just requires a difficulty choice to be passed to the questionGenerator function and then we can return the question as a string.
We could also extend this by using eval() to get the answer as well, but eval() is bad practice in general. Any thoughts on this feel free to let me know